Stateless allocator support for formatting
Classes are parameterized with allocator type as needed for the allocator to be
used to allocate formatting buffer space only. Many other allocating sites are
not covered yet, for instance, clone() and create() functions, exceptions,
creation of internal objects like formatters, and more.

Default logger, default registry, and other facilities provided by global
(namespace) functions are not covered and work as before. For classes that are
affected, legacy behaviour can be achieved by using std::allocator as an
argument. Many classes (that appear to be user facing) have the default
allocator parameter specified as std::allocator. For even more seamless upgrade
experience, some templatized classes are renamed using basic_ prefix, and an
alias with the old name provided.

Newly templatized classes have instantiation with std::allocator added to the
cpps, so that it's possible to use them as a shared lib. In order to use spdlog
with other allocators, it has to be used as header only.

Only stateless allocators are supported by this change, i.e. there is no way now
to pass an instance of polymorphic_allocator to spdlog objects.