Fixed: memory_buffer on nonpropagating alloc in container
The implementation of basic_memory_buffer::move() always propagated the underlying allocator, however with non-propagating allocators such as polymorphic_allocator that is not correct and doesn't compile, because it has all assignment operators deleted.

This change detects if the allocator should propagate on move assignment and retains the behaviour if it should. Otherwise the expected behaviour should depend on whether the target's allocator is the same. If it is, the data is just moved over as usual. If it is not, it is expected that the container falls back to copying. However since memory_buffer does not copy, the code throws a logic_error.

This change allows  to store fmt::basic_memory_buffer<T,S,std::pmr::polymorphic_allocator> in containers such as std::pmr::vector, provided that both the vector and the elements placed into it use the same allocator instance.