Stateful allocator support for formatting
Allocator objects are added to logger, formatter, and sink classes. They
are initialized by the last parameter of class ctors, and the classes
define `allocator_type` alias to conform to uses-allocator construction
protocol.

The allocator objects are passed into fmt::memory_buffer objects so that
they will use proper allocators while formatting. Other allocating
entities in the library are not in the scope and still use new/delete
allocation.