Stateful allocator support for formatting
Allocator objects are added to logger, formatter, and sink classes. They are
initialized by the last parameter of class ctors, and the classes define
`allocator_type` alias to conform to uses-allocator construction protocol.

The allocator objects are passed into fmt::memory_buffer objects so that they
will use proper allocators while formatting. Other allocating entities in the
library are not in the scope and still use new/delete allocation.

Classes that contain allocator objects do that by inheriting from the allocator
in order to benefit from empty base optimization: in the case of stateless
allocators, the classes object representation size will be the same as before
the allocator were introduced into spdlog.

Unversal copy/move assignment (previously done as swap) is replaced with
proper implementations because swap-to-assign does not work well with
allocators.