Allocator aware logger name
Logger name is stored in basic_logger and in registry using the custom allocator type. For stateful allocators, an allocator instance different from the one used for formatting buffers can be specified in the logger ctors. That other instance will later be used for other allocations like vector of sinks, etc. The former allocator instance renames to "formatting buffer allocator".